+++ src/display/mod.rs
@@ -1,2 +1,5 @@
-pub mod display_graph;
-pub mod visualizer;
+mod display_graph;
+mod visualizer;
+
+pub use display_graph::*;
+pub use visualizer::*;
