+++ src/utils/mod.rs
@@ -1,2 +1,5 @@
-pub mod DisjointUnionFind;
-pub mod graph;
+mod disjointunionfind;
+mod graph;
+
+pub use disjointunionfind::*;
+pub use graph::*;
